pyproject.toml
[project]
name = "quant-trading-system"
version = "0.1.0"
description = "Quantitative trading platform — pragmatic single-process architecture"
requires-python = ">=3.12"
dependencies = [
    # Core
    "pydantic>=2.0,<3.0",
    "pydantic-settings>=2.0,<3.0",
    # Data
    "pandas>=2.0,<3.0",
    "numpy>=1.26,<3.0",
    "yfinance>=0.2,<1.0",
    # Database
    "sqlalchemy>=2.0,<3.0",
    "psycopg2-binary>=2.9,<3.0",
    "alembic>=1.13,<2.0",
    # API
    "fastapi>=0.110,<1.0",
    "uvicorn[standard]>=0.27,<1.0",
    "websockets>=12.0,<14.0",
    "python-jose[cryptography]>=3.3,<4.0",
    # Optimization
    "cvxpy>=1.4,<2.0",
    "scipy>=1.12,<2.0",
    # CLI
    "typer>=0.12,<1.0",
    "rich>=13.0,<14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-asyncio>=0.23,<1.0",
    "httpx>=0.27,<1.0",
    "ruff>=0.3,<1.0",
    "mypy>=1.8,<2.0",
]

[project.scripts]
quant = "src.cli.main:app"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.12"
strict = true
